refactor: move MenuInput and DBInput to Input package
